project/ossia-user-service/pyproject.toml
[project]
name = "ossia-service-users"
version = "1.0.0"
description = "Users service"
authors = [
    { name = "justxale" }
]
requires-python = ">=3.12,<4.0"
readme = "README.md"
dependencies = [
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "faststream[rabbit] (>=0.5.42,<0.6.0)",
    "granian>=2.5.0",
    "orjson (>=3.10.18,<3.11.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<2.11.0)",
    "python-multipart>=0.0.20",
    "tortoise-orm[accel,asyncpg] (>=0.25.1,<0.26.0)",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.9.4,<0.10",
]
prod = [
    "granian[uvloop]>=2.5.0",
    "uvloop>=0.21.0",
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
addopts = '-v'
